requirements.txt
robotframework==4.1.2
robotframework-pageobjectlibrary==1.1
robotframework-seleniumlibrary==5.1.3
selenium==4.0.0
